Made QueueItems clean up better, and a deletion feature - both on demand and after X days

+++ mass-pnr-lookup/Queues/ExtractionQueue.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Web;
 using CprBroker.Engine.Queues;
 using mass_pnr_lookup.Models;
+using CprBroker.Engine.Local;
 
 namespace mass_pnr_lookup.Queues
 {
@@ -17,33 +18,46 @@ namespace mass_pnr_lookup.Queues
                 using (var context = new BatchContext())
                 {
                     var batch = context.Batches.Find(item.BatchId);
-                    var parser = batch.CreateParser();
 
-                    try
+                    if(batch == null)
                     {
-                        // Delete existing lines if needed
-                        var oldLines = context.BatchLines.Where(bl => bl.Batch_BatchId == batch.BatchId);
-                        context.BatchLines.RemoveRange(oldLines);
-                        context.SaveChanges();
-
-                        context.Entry<Batch>(batch).Reload();
-
-                        var lines = parser.ToArray();
-                        if (batch.Lines == null)
-                            batch.Lines = new List<BatchLine>();
-
-                        Array.ForEach<BatchLine>(lines, bl => batch.Lines.Add(bl));
-
-                        batch.EnqueueAllAfterExtraction(context);
+                        // This happens if the batch is deleted before extraction
+                        ret.Add(item);
                     }
-                    catch (Exception ex)
+                    else
                     {
-                        Logger.LogException(ex);
+                        try
+                        {
+                            var parser = batch.CreateParser();
+                            // Delete existing lines if needed
+                            var oldLines = context.BatchLines.Where(bl => bl.Batch_BatchId == batch.BatchId);
+                            context.BatchLines.RemoveRange(oldLines);
+                            context.SaveChanges();
 
-                        batch.Status = BatchStatus.Error;
-                        context.SaveChanges();
+                            context.Entry<Batch>(batch).Reload();
+
+                            var lines = parser.ToArray();
+                            if (batch.Lines == null)
+                                batch.Lines = new List<BatchLine>();
+
+                            Array.ForEach<BatchLine>(lines, bl => batch.Lines.Add(bl));
+
+                            batch.EnqueueAllAfterExtraction(context);
+                            ret.Add(item);
+                        }
+                        catch (Exception ex)
+                        {
+                            Admin.LogException(ex);
+
+                            batch.Status = BatchStatus.Error;
+                            context.SaveChanges();
+                        }
+                        if (item.Impl.AttemptCount >= this.Impl.MaxRetry - 1)
+                        {
+                            // Max retry reached, clean up queueItem
+                            ret.Add(item);
+                        }
                     }
-                    ret.Add(item);
                 }
             }
             return ret.ToArray();
